Add preliminary support for Launchpad S
The Launchpad S doesn't have RGB LEDs, it only has red and green,
with 4 levels each. This commit detects a Launchpad S, making it
usable with Griode (with the proper mapping for the 64 pads and
the 8 essential buttons) but it doesn't map the colors at this
point.

This is expected to be useful if somebody wants to add support for
the Launchpad Mini (because I think it is closer to the Launchpad S).