Refactor the use of colors to better handle Launchpad Mini
The Launchpad Mini (and the Launchpad S) has a limited palette.
If we send it the same color codes as the Launchpad RGB / Pro,
it doesn't look good, and many colors are difficult to contrast.

This adds an intermediary "palette" layer. Now, in the code,
instead of colors.RED we should use palette.CHANNEL[0] (for
instance). See palette.py for more details.

At this point, it is still possible to use colors.X, and it
will generate warnings. I will continue to replace colors.X
with palette.Y to finish the refactoring.