New step sequencer
The original "looper" was a good proof of concept, but it had many
shortcomings (including a pretty horrible user interface).

This is another attempt, centered around a step sequencer.
This is the MVP, allowing to record simple rythmic patterns.
It probably has a bunch of bugs, and needs to be improved,
but it's functional enough for simple demos so I'm commiting it.